aviation-rag-system/requirements.txt
# Aviation RAG System Dependencies

# Core ML and NLP
torch>=2.0.0
sentence-transformers>=2.2.2
numpy>=1.21.0

# PDF Processing
PyPDF2>=3.0.1
pdfplumber>=0.9.0

# Database
pymongo>=4.5.0

# Claude API
anthropic>=0.34.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.0.0

# Logging and Monitoring
structlog>=23.0.0
prometheus-client>=0.17.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Optional: Advanced features
scikit-learn>=1.3.0  # For additional similarity metrics
nltk>=3.8.0  # For advanced text processing
spacy>=3.7.0  # For NLP tasks

# Optional: Monitoring and observability
sentry-sdk[fastapi]>=1.32.0  # Error tracking